Resolve "Create an API call, with which to answer questions as a lecturer."

Closes #40

See merge request cse1105/2020-2021/team-repositories/oopp-group-54/repository-template!35

+++ server/src/main/java/nl/tudelft/oopp/group54/controllers/answers/AnswerService.java
@@ -6,8 +6,8 @@ import java.util.Map;
 
 public interface AnswerService {
 
-    Map<String, Object> answerQuestion(Long lectureId,
+    Map<String, Object> answerQuestion(Integer lectureId,
                                        String userId,
-                                       Long questionId,
+                                       Integer questionId,
                                        @Nullable String answerText);
 }
